Changed "exg" instruction to "exy" & and restored "exg" with a smarter macro.
Clownacy's 2017 patch to ASM68K that eliminated register reordering on the "exg" instruction had a nasty unintentional side effect that produces illegal encoding when an address register is listed as the first operand and a data register is listed as the second operand. This is because, when exchanging between register types, the first register field in the exg instruction represents the data register, while the second one represents the address register.

So, while "exg dn,dn", "exg an,an", and "exg dn,an" are all valid formats as-is, "exg an,dn" will not encode correctly by default and needs to be reordered. The way ASM68K originally handled this was by making the whichever register has the lowest internal ID number the first operand, while the other is made the second operand. However, as this behavior has been NOP'd out, there is now no internal way for this reordering to be ensured when it's necessary for the instruction to be encoded correctly.

As such, I have changed the internal instruction name to "exy" and created a macro that will reorder the registers when an address register is listed first and a data register is listed second; this reordering will not be applied to any other case, allowing for things like the Sonic 1 Disassembly to still be built to a bit-perfect spec.